pyproject.toml
[tool.poetry]
name = "evalbot"
version = "1.0.0"
description = "A small discord bot to demo python 3.10 features."
authors = ["Hassan Abouelela <hassan@hassanamr.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.*"
"discord.py" = "~=1.7.3"
sentry-sdk = "~=1.3"
coloredlogs = "~=14.0"

[tool.poetry.dev-dependencies]
python-dotenv = "~=0.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
